kernel: move FilenameCache to io.c, use for INPUT_FILENAME
Previously, INPUT_FILENAME() returned a new (mutable) string object each
time it was called. Now, we always return the same (immutable) string
object from the cache of input filenames we keep anyway (for profiling,
amongst other things). Note that this string is used for tracking the
locations where we declare operations or install methods, so the memory
used for the duplicate strings could add up considerably before this
patch.
